+++ ops/deploy.mjs
@@ -1,39 +1,26 @@
 import mime from 'mime-types';
-import S3SyncClient from 's3-sync-client';
-import AWS from "aws-sdk";
+import { S3Client } from '@aws-sdk/client-s3';
+import { S3SyncClient } from 's3-sync-client';
 import * as fs from 'fs';
 
 // Read config file
 let config = JSON.parse(fs.readFileSync('../config.json', 'utf-8'));
 
-AWS.config.getCredentials(function(err) {
-  if (err) {
-      console.log('credentials not loaded');
-      console.log(err.stack);
-      process.exit(1);
-  } else {
-    console.log("Access key:", AWS.config.credentials.accessKeyId);
-  }
-});
-
 const run = async () => {
     try {
-        const sync = new S3SyncClient({
-            region: config.aws.region,
-            credentials: {
-                accessKeyId: AWS.config.credentials.accessKeyId,
-                secretAccessKey: AWS.config.credentials.secretAccessKey,
-            }
+        const s3Client = new S3Client({
+            region: config.aws.region
+        });
+        const { sync } = new S3SyncClient({
+            client: s3Client
         });
 
-        console.log(`Copying files with ACL and content-type...`);
-        await sync.bucketWithLocal('../build/', config.aws.site_bucket, {
-            commandInput: {
+        console.log(`Syncing files...`);
+        await sync('../build/', config.aws.site_bucket, {
+            commandInput: (input) => ({
                 ACL: "public-read",
-                ContentType: (syncCommandInput) => (
-                    mime.lookup(syncCommandInput.Key) || 'text/html'
-                ),
-            },
+                ContentType: mime.lookup(input.Key) || 'text/html'
+            }),
             del: true 
         });
         console.log(`Done!`);