+++ ops/invalidate.mjs
@@ -1,23 +1,15 @@
-import AWS from "aws-sdk";
+import { CloudFrontClient, CreateInvalidationCommand } from '@aws-sdk/client-cloudfront';
 import * as fs from 'fs';
 
 // Read config file
 let config = JSON.parse(fs.readFileSync('../config.json', 'utf-8'));
 
-AWS.config.getCredentials(function(err) {
-  if (err) {
-      console.log('credentials not loaded');
-      console.log(err.stack);
-      process.exit(1);
-  } else {
-    console.log("Access key:", AWS.config.credentials.accessKeyId);
-  }
-});
-
 const run = async () => {
     try {
-        var cloudfront = new AWS.CloudFront();
-        var params = {
+        var cloudfront = new CloudFrontClient({
+          region: config.aws.region
+        });
+        var input = {
           DistributionId: config.aws.cloudfront_distribution,
           InvalidationBatch: {
             CallerReference: new Date().toString(),
@@ -29,10 +21,10 @@ const run = async () => {
             }
           }
         };
-        cloudfront.createInvalidation(params, function(err, data) {
-          if (err) console.log(err, err.stack); // an error occurred
-          else     console.log(data);           // successful response
-        });
+
+        const command = new CreateInvalidationCommand(input);
+        const response = await cloudfront.send(command);
+        console.log(response);
         console.log(`Done!`);
     } catch (err) {
         console.log("Error", err);