+++ cms/markdown/about/changelog.md
@@ -1,3 +1,7 @@
 
+#### 2025-02-16
+* Minor loading improvements
+* Build process changed from Snowpack to Vite
+* Adding test workflow
 #### 2021-08-19
 * Improved styling for readability and clarity